Split to multiple helpers to separate concerns
